hw/intc/arm_gic_kvm: set up a flat irqchip routing table
In case the kernel supports GSI routing, we set up
irqchip routing entries for GIC SPIs. This is not mandated
as long as MSI routing is not used (because the kernel sets
a default irqchip routing table). However once MSI routing gets
used (for VIRTIO-PCI vhost for example), the first call to
KVM_SET_GSI_ROUTING overrides the kernel default irqchip table.
If no routing entry exists for the GSI, any IRQFD signaling for
this GSI will fail.

Signed-off-by: Eric Auger <eric.auger@linaro.org>

---

gsi routing for ARM is functional but needs to be maturated.
The API for MSI routing still is in discussion.
See http://www.spinics.net/lists/kvm/msg118792.html