hw/intc/arm_gicv3_its: Implement reset
Currently the ITS is not reset and this causes trouble
when state backup is initiated before the guest has initialized
the ITS registers and especially GITS_CBASER<n>.

We are likely to save register values set before the reset/
restart. The register values may not be consistent with the
data structures in RAM.

So let's use the ITS KVM device new combo,
KVM_DEV_ARM_VGIC_GRP_CTRL/KVM_DEV_ARM_ITS_CTRL_RESET
to explicitly force the in-kernel emulated reset.

Signed-off-by: Eric Auger <eric.auger@redhat.com>