hw/arm/virt-acpi-build: Add virtio-iommu node in IORT table
This patch builds the virtio-iommu node in the ACPI IORT table.

The RID space of the root complex, which spans 0x0-0x10000
maps to streamid space 0x0-0x10000 in the virtio-iommu which in
turn maps to deviceid space 0x0-0x10000 in the ITS group.

The iommu RID (0x8) is excluded as described in virtio-iommu
specification.

Signed-off-by: Eric Auger <eric.auger@redhat.com>

---

v7 -> v8:
- exclude the iommu RID (0x8) in the root complex ID mapping