qdev: export qbus_find_recursive
We intend to create the virtio-iommu from the virt machine
realize() code and not by passing the -device virtio-iommu-device
option. That way the instantiation of the virtio-iommu can
depend on a virt machine option, as it is currently done for
the smmuv3 iommu.

Its parent bus, the virtio-bus will need to be located and
set. We intend to locate it by using qbus_find_recursive().

Signed-off-by: Eric Auger <eric.auger@redhat.com>